Move full ConfigureRoutes()
